// src/app/services/enums/pageenum.ts
export enum Page {
    Morning = 'morning',
    Afternoon = 'afternoon',
    Night = 'night',
    CultureGuide = 'morning_culture-guide',
    BrandManual = 'morning_brand-manual',
    Compass = 'morning_compass',
    Tree = 'morning_tree',
    Guardians = 'morning_guardians',
    Magazine = 'morning_magazine',
    Colab = 'morning_colab',
    Us = 'night_us',
    Stamp = 'night_stamp',
    Network = 'night_network',
    Talks = 'night_talks',
    Gabw = 'https://www.gabw.cc',
    Contact = 'night_contact',
    Timeline = 'afternoon_time-line'
}